+++ CHANGELOG.md
@@ -7,6 +7,10 @@ A preview of the next release can be installed from
 
 [Babashka](https://github.com/babashka/babashka): Native, fast starting Clojure interpreter for scripting
 
+## Unreleased
+
+- Fix `with-redefs` + `intern` (see SCI issue [#973](https://github.com/babashka/sci/issues/973)_
+
 ## 1.12.202 (2025-06-15)
 
 - Support `clojure.lang.Var/{get,clone,reset}ThreadBindingFrame` for JVM Clojure compatibility